Well core/lib.rs is simple now
